improved authorization

+++ CertificateManager.Logic/RoleManagementLogic.cs
@@ -1,19 +1,28 @@
 ﻿using CertificateManager.Entities;
 using CertificateManager.Entities.Exceptions;
+using CertificateManager.Logic.Interfaces;
 using CertificateManager.Repository;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 
 namespace CertificateManager.Logic
 {
     public class RoleManagementLogic
     {
-        IConfigurationRepository configurationRepository;
+        private const string wellKnownAdministratorRoleId = "092020ed-57cf-4af6-a1f6-6e1fdd7f5e8a";
 
-        public RoleManagementLogic(IConfigurationRepository configurationRepository)
+        public static Guid WellKnownAdministratorRoleId { get { return new Guid(wellKnownAdministratorRoleId); } }
+        public static List<Guid> DefaultTemplateIssuerRoles { get { return new List<Guid>() { WellKnownAdministratorRoleId }; } }
+
+        IConfigurationRepository configurationRepository;
+        IAuthorizationLogic authorizationLogic;
+
+        public RoleManagementLogic(IConfigurationRepository configurationRepository, IAuthorizationLogic authorizationLogic)
         {
             this.configurationRepository = configurationRepository;
+            this.authorizationLogic = authorizationLogic;
         }
 
         public SecurityRole GetRole(Guid id)
@@ -26,19 +35,25 @@ namespace CertificateManager.Logic
             return configurationRepository.GetSecurityRoles();
         }
 
-        public SecurityRole AddRole(SecurityRole entity)
+        public SecurityRole AddRole(SecurityRole entity, ClaimsPrincipal user)
         {
             entity.Id = Guid.NewGuid();
             configurationRepository.InsertSecurityRole(entity);
             return entity;
         }
 
-        public void DeleteRole(SecurityRole securityRole)
+        public void DeleteRole(SecurityRole securityRole, ClaimsPrincipal user)
         {
+            if (authorizationLogic.AuthorizedToManageRoles(user))
+                throw new UnauthorizedAccessException("Current user context is not authorized to manage roles");
+
+            if (securityRole.Id == RoleManagementLogic.WellKnownAdministratorRoleId)
+                throw new InvalidOperationException("The built-in administrator account cannot be deleted");
+
             configurationRepository.DeleteSecurityRole(securityRole);
         }
 
-        public void UpdateRole(SecurityRole securityRole)
+        public void UpdateRole(SecurityRole securityRole, ClaimsPrincipal user)
         {
             configurationRepository.UpdateSecurityRole(securityRole);
         }
@@ -60,7 +75,7 @@ namespace CertificateManager.Logic
             List<AuthenticablePrincipal> members = new List<AuthenticablePrincipal>();
             foreach (Guid memberId in role.Member)
             {
-                members.Add(configurationRepository.GetAuthenticablePrincipal(memberId));
+                members.Add(configurationRepository.GetAuthenticablePrincipal<AuthenticablePrincipal>(memberId));
             }
 
             return members;
@@ -75,7 +90,7 @@ namespace CertificateManager.Logic
 
         public AuthenticablePrincipal AddRoleMember(Guid roleId, Guid memberId)
         {
-            AuthenticablePrincipal principal = configurationRepository.GetAuthenticablePrincipal(memberId);
+            AuthenticablePrincipal principal = configurationRepository.GetAuthenticablePrincipal<AuthenticablePrincipal>(memberId);
 
             if (principal == null)
                 throw new ReferencedObjectDoesNotExistException("Specified member id was not found. No changes have been made.");
@@ -94,16 +109,49 @@ namespace CertificateManager.Logic
             return principal;
         }
 
+        public SecurityRole GetCertificateManagerAdministrator()
+        {
+            return configurationRepository.GetSecurityRole(WellKnownAdministratorRoleId);
+        }
+
+        public void SetRoleScopes(Guid roleId, List<Guid> scopes, ClaimsPrincipal user)
+        {
+            if (!authorizationLogic.IsAuthorized(AuthorizationScopes.ManageRolesScope, user))
+                throw new UnauthorizedAccessException("Access denied: scope change will not be performed.");
+
+            SecurityRole role = configurationRepository.GetSecurityRole(roleId);
+
+            List<Scope> validScopes = authorizationLogic.GetAvailibleScopes();
+            if(scopes != null && scopes.Any() )
+            {
+                foreach (Guid scope in scopes)
+                {
+                    if (!validScopes.Select(item => item.Id).Contains(scope))
+                        throw new ReferencedObjectDoesNotExistException("Requested scope does not exist");
+                }
+            }
+            else
+            {
+                scopes = new List<Guid>();
+            }
+
+            role.Scopes = scopes;
+
+            configurationRepository.UpdateSecurityRole(role);
+        }
+
         public void InitializeRoles(List<Guid> users)
         {
             SecurityRole adminRole = new SecurityRole()
             {
                 Enabled = true,
-                Id = Guid.NewGuid(),
+                Id = WellKnownAdministratorRoleId,
                 Member = users,
-                Name = "Administrators@certificatemanager.local"
+                Name = "Administrators@certificatemanager.local",
+                Scopes = authorizationLogic.GetAvailibleScopes().Select(scope => scope.Id).ToList()
             };
 
+            configurationRepository.InsertSecurityRole(adminRole);
         }
     }
